+++ FE/.gitignore
@@ -1,4 +1,5 @@
 package-lock.json
+test.json
 
 # Logs
 logs
